+++ packages/qfai/src/core/index.ts
@@ -1,6 +1,7 @@
 export * from "./config.js";
 export * from "./decisionGuardrails.js";
 export * from "./ids.js";
+export * from "./preflight/sddPreflight.js";
 export * from "./report.js";
 export * from "./types.js";
 export * from "./validate.js";